correct CRC caculate\r\n tested with \r\n\t:10246200464C5549442050524F46494C4500464C33# \r\n\t:02000004FFFFFC# \r\n\t:020000021200EA#

+++ Inc/hex_record.h
@@ -10,7 +10,7 @@ uint8_t hex_digit(uint8_t xChar);
 uint8_t char_digit(uint8_t xHex);
 
 uint8_t record_get_u8(const uint8_t *pBuffer);
-uint8_t record_get_u16(const uint8_t *pBuffer);
+uint16_t record_get_u16(const uint8_t *pBuffer);
 
 uint16_t record_set_u8(uint8_t *pBuffer, uint8_t u8Data);
 uint16_t record_set_u16(uint8_t *pBuffer, uint16_t u16Data);
